refactor(persistence): split repositories.py into per-domain protocol files (#1557)
