+++ src/synthorg/settings/service.py
@@ -38,7 +38,7 @@ from synthorg.settings.models import SettingDefinition, SettingEntry, SettingVal
 
 if TYPE_CHECKING:
     from synthorg.communication.bus_protocol import MessageBus
-    from synthorg.persistence.repositories import SettingsRepository
+    from synthorg.persistence.settings_protocol import SettingsRepository
     from synthorg.settings.encryption import SettingsEncryptor
     from synthorg.settings.registry import SettingsRegistry
