+++ tests/unit/persistence/sqlite/test_decision_repo.py
@@ -8,8 +8,8 @@ import pytest
 
 from synthorg.core.enums import DecisionOutcome
 from synthorg.engine.decisions import DecisionRecord
+from synthorg.persistence.decision_protocol import DecisionRepository
 from synthorg.persistence.errors import DuplicateRecordError, QueryError
-from synthorg.persistence.repositories import DecisionRepository
 from synthorg.persistence.sqlite.decision_repo import SQLiteDecisionRepository
 
 if TYPE_CHECKING:
