+++ tests/unit/persistence/test_protocol.py
@@ -14,31 +14,33 @@ from synthorg.hr.persistence_protocol import (
     LifecycleEventRepository,
     TaskMetricRepository,
 )
+from synthorg.persistence.agent_state_protocol import AgentStateRepository
+from synthorg.persistence.artifact_protocol import ArtifactRepository
+from synthorg.persistence.audit_protocol import AuditRepository
+from synthorg.persistence.checkpoint_protocol import (
+    CheckpointRepository,
+    HeartbeatRepository,
+)
+from synthorg.persistence.cost_record_protocol import CostRecordRepository
+from synthorg.persistence.decision_protocol import DecisionRepository
+from synthorg.persistence.message_protocol import MessageRepository
+from synthorg.persistence.parked_context_protocol import ParkedContextRepository
 from synthorg.persistence.preset_repository import (
     PersonalityPresetRepository,
     PresetListRow,
     PresetRow,
 )
+from synthorg.persistence.project_protocol import ProjectRepository
 from synthorg.persistence.protocol import PersistenceBackend
-from synthorg.persistence.repositories import (
-    AgentStateRepository,
-    ApiKeyRepository,
-    ArtifactRepository,
-    AuditRepository,
-    CheckpointRepository,
-    CostRecordRepository,
-    DecisionRepository,
-    HeartbeatRepository,
-    MessageRepository,
-    ParkedContextRepository,
-    ProjectRepository,
-    SettingsRepository,
-    TaskRepository,
-    UserRepository,
-)
+from synthorg.persistence.settings_protocol import SettingsRepository
+from synthorg.persistence.task_protocol import TaskRepository
 from synthorg.persistence.training_repos import (
     TrainingPlanRepository,
     TrainingResultRepository,
+)
+from synthorg.persistence.user_protocol import (
+    ApiKeyRepository,
+    UserRepository,
 )
 from synthorg.persistence.workflow_definition_repo import WorkflowDefinitionRepository
 from synthorg.persistence.workflow_execution_repo import WorkflowExecutionRepository
