+++ tests/unit/settings/test_service_set_many.py
@@ -12,7 +12,7 @@ from cryptography.fernet import Fernet
 from pydantic import BaseModel, ConfigDict
 
 from synthorg.api.errors import VersionConflictError
-from synthorg.persistence.repositories import SettingsRepository
+from synthorg.persistence.settings_protocol import SettingsRepository
 from synthorg.settings.encryption import SettingsEncryptor
 from synthorg.settings.enums import SettingNamespace, SettingType
 from synthorg.settings.errors import (
